backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
PyJWT[crypto]==2.10.1
cryptography==44.0.0
httpx==0.28.1
psycopg[binary]==3.2.3
pytest==8.3.4
pytest-asyncio==0.24.0
